JavaScript Math Game
This was a practice project from July, 2022. The purpose of this project was to make a simple web-based math game using JavaScript. To summarize the gameplay, the user will have the opportunity to read a brief description of the game at the top of the web page before playing the game. The user will then have to click the 'Start' button in order to play the game. A random addition expression will be generated once the user clicks the 'Start' button and will have 10 seconds to solve the given expression by entering their answer into a textbox and clicking the 'Submit' button. The user will have to do this 10 times before the game ends and displays the user's total score out of 10 questions. In the case that the user does not enter an answer within 10 seconds, the game will simply move on to the next question if one is available. Once the game is over, the user will have to reset the game before being able to play again.