Practice Online Store Hompage (HTML, CSS, jQuery)
This was a practice project from April, 2022 for the purpose of strengthening my application of HTML, CSS and jQuery, specifically jQuery animation. I practiced the application of these languages by creating a home page for a fictional street fashion shoe store. In its entirety, you can expect to see jQuery and CSS animation as well as dynamic and static CSS styling of HTML elements. All images used in this project are for practice purposes only; I do not own these images and did not nor do I intend to use this web page for commercial purposes.