JavaScript Math Game
This project was created in July, 2022. The purpose of this project was to create a simple web-based game using JavaScript. To summarize the gameplay of this game, the user will have the opportunity to read the brief description of the game provided at the top of the webpage. Once the user is comfortable with their understanding of the game, the user will then have to click the 'Start' button to start the game. Once the 'Start' button is clicked, they will be met with a random addition expression in which they have 10 seconds to answer. Users will enter their answers into the textbox below the expression and click the 'Submit' button to the right of the textbox to submit their answer. If the user does not enter an answer within 10 seconds, the game will move on to the next question. The user will be tasked with answering 10 questions before the game ends and relays their score out of 10 questions. If the user wishes to play again, they will have to click the 'Reset' button.