Customizable JavaScript Calculator
This project was created in January, 2022. The purpose of this project was to create a fully functional web-based graphical calculator that users could customize the colour of with the available colour palette provided. I also wanted to implement a calculation history log that collects user calculations and makes them both readable and voidable depending on a user's wishes.