jQuery Search Engine
This project was created in April, 2022. The purpose of this project was to create a limited search engine web application using jQuery that essentially manipulates the 'href' attribute of an anchor tag that encapsulates an HTML button element. When this button is clicked, it will redirect the user to the specified domain name inputted by the user in the textbox provided. Based on the selected extension specified by the user in the select menu located in the top left corner of the web page, the user will be able to access different versions of a given domain name using the following extensions: .ca, .com and .uk.