Website Diagnostic Tool for Linux
This project was created in October, 2022. The purpose of this assignment was to have the nslookup, ping and traceroute results of one or more websites to be stored in an HTML file that could then be accessed by the user and examined. A key objective of this project was to use bash scripting to generate HTML markups containing bash command output. This script only works for machines running Linux based operating systems with dnsutils and traceroute installed onto the machine. If you are having problems running this code, it may be due to the fact that dnsutils and traceroute have not been installed. To install these packages, open a terminal window on your Linux machine and type in the following commands:

sudo apt install dnsutils
sudo apt install traceroute