JavaScript ATM Web Application
This project was created in the month of July of the year 2022. The purpose of this project was to create an application that mimics the ATM cash withdrawal menu of Canadian ATM's using JavaScript, HTML and CSS. The application upon loading or refreshing will generate a random account balance for the user. Users will be tasked with entering the amount of money they wish to withdraw from their account balance by using the on-screen keypad provided to them. Once they do so, they will be tasked with specifying the amount of 5, 10, 20, 50 and/or 100 dollar bills they wish the application to give them to match their requested withdrawal amount. Finally, the user will click on the 'Withdraw Funds' button to withdraw their requested amount from their account balance and a receipt of their withdrawal as well as their current account balance will be presented to them. If the requested withdrawal amount exceeds what is present in their account balance, the application will ask the user to lower their withdrawal amount to be equal or less than their account balance. Also, if the selection of bills requested by the user exceeds or is lower than the requested withdrawal amount inputted, the application will alert the user to alter their bill selection to match the requested withdrawal amount.