JavaScript On-Screen Keyboard Application
This project was created in the month of July of the year 2022. The purpose of this application was to create a web based on-screen keyboard that was hacker themed using HTML, CSS and JavaScript. This application allows users to write complete sentences with punctuation and special characters as well as change the case of their letter input using the 'Shift' and 'Caps' buttons of the on-screen keyboard. Users can even format their text using the 'Tab' button to indent text and the 'Enter' button to move to a new line. Numerical input is also supported by this on-screen keyboard through the use of a number pad. Using CSS and HTML, I was able to implement the QWERTY key format of conventional keyboards, replicate a typical number pad containing all ten digits from 0-9, designate punctuation and special character inputs to their own section for ease of access, and style the webpage to have a classic hacker aesthetic with a black and lime green colour palette .