4Chan Homepage Replica
This project was created in the month of October of the year 2021. The purpose of this project was to use the HTML and CSS code I knew at the time to replicate the homepage of the famous web forum 4Chan by sight alone. You can reference https://www.4chan.org/ to compare my version of the 4Chan homepage to the official 4Chan homepage.