Facebook Login Page Replica
This project was created in the month of October of the year 2021. The purpose of this project was to use the HTML and CSS code I knew at the time to replicate the login page of the famous social media website Facebook by sight alone. You can reference https://www.facebook.com/login/ to compare my version of the Facebook login page to the official Facebook login page.