Twitter Signup Page Replica
This project was created in the month of October of the year 2021. The purpose of this project was to use the HTML and CSS code I knew at the time to replicate the signup page of the famous social media website Twitter by sight alone. Since Twitter's signup page is no longer the way it was a year ago, you will be unable to compare the signup page I made to the signup page officially made by Twitter's development team to test for accurate replication on my behalf.